+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 0.4.2
+
+- **Added** iterator over ranges, `iter_from` method and `range` function, indexing arrays by `Ranged`.
+- **Added** [sudoku](examples/sudoku.rs) example.
+
 ## 0.4.1
 
 - **Added** to and from `isize` and `usize` conversions